Claude/requirements.txt
anthropic
feedparser
beautifulsoup4
python-dotenv
